### Summary

This Pull Request introduces major refactoring to standardize primary key identifiers across the entire application stack and includes necessary fixes for authentication providers and CORS configuration.

The primary goal is to standardize all internal JPA Entity and DTO identifiers to use the `Long` data type, preparing the application for future scaling and consistency.

### Key Changes and Refactors

* **Standardization to `Long` IDs:**
    * Updated all JPA entities to use `Long` for primary keys and optimized Lombok annotations for cleaner code.
    * Adapted JPA repositories to correctly handle and query with `Long` IDs.
    * Refactored DTOs and the service layer logic to correctly process and pass `Long` identifiers throughout the application.
* **Performance and Security:**
    * Optimized `AuthUtils` to reuse the existing security context for improved performance.
    * Unified the global error response structure to provide consistent feedback and include the requested API path in error details.
---

### Fixes and Configuration

* **Auth:** Resolved an issue causing duplicate entries or conflicts with authentication providers.
* **CORS:** Explicitly configured CORS settings to allow necessary traffic from the frontend application domain.
* **Database Queries:** Cleaned up and optimized `UsuarioRepository` custom queries and updated consuming services for better clarity and efficiency.
* Minor datatype fixes applied across various files.

+++ server/src/main/java/com/xtheggx/monedaDOS/controller/CategoriaController.java
@@ -38,7 +38,7 @@ public class CategoriaController {
 
     // 3. Actualizar categoría
     @PutMapping("/{id}")
-    public ResponseEntity<?> actualizar(@PathVariable Integer id, @Valid @RequestBody CategoriaDTO dto) {
+    public ResponseEntity<?> actualizar(@PathVariable Long id, @Valid @RequestBody CategoriaDTO dto) {
         Long userId = authUtils.currentUserId();
         Categoria actualizada = categoriaService.actualizarCategoria(id, userId, dto);
         return ResponseEntity.ok(actualizada);
@@ -46,7 +46,7 @@ public class CategoriaController {
 
     // 4. Eliminar categoría
     @DeleteMapping("/{id}")
-    public ResponseEntity<?> eliminar(@PathVariable Integer id) {
+    public ResponseEntity<?> eliminar(@PathVariable Long id) {
         Long userId = authUtils.currentUserId();
         categoriaService.eliminarCategoria(id, userId);
         return ResponseEntity.ok(Map.of("mensaje", "Categoría eliminada"));
